Big_Mamma/requirements.txt
pandas==1.4.2
numpy==1.22.3
requests==2.27.1
google-api-python-client==2.46.0
